fix: pass correct repo url for nightly
The amendMatrixForNext function uses the repository argument to
determine the version constraint for the returned matrix for kind
nightly.
Previously the upstream-mirror repo url was passed as an argument, but
the nightly builds use a different repo url.
This resulted in failed builds, because the version string 'next' could
not be parsed by composer.
Example:
https://github.com/mage-os/generate-mirror-repo-js/actions/runs/6370219504/job/17291152703

With this change, the nightly kind will always return the `@alpha`
version constraint, that is, stability constraint.
There currently is no way to distinguish between different nightly
repos, but since both work the same way, this is fine (for now).